Fix cleaning up of database files in tests.

+++ test/helper.py
@@ -16,13 +16,14 @@ class AbstractTest(object):
         self.sess.close()
         try:
             os.unlink(self.cfg.sqldb)
-        except:
+        except OSError:
             pass
 
-        try:
-            os.unlink(self.cfg.plist)
-        except:
-            pass
+        if self.cfg.plist is not None:
+            try:
+                os.unlink(self.cfg.plist)
+            except OSError:
+                pass
 
     def _read_in_plist(self):
         """ returns a list of filenames (sorted) """
